requirements.txt
pynput>=1.7.6
numpy
scikit-learn
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
pystray>=0.19.0
Pillow>=8.0.0
pywin32>=300
psutil>=5.4.0
